extract login into a separate component
